requirements.txt
evaluate 
bert-score 
pandas
python-dotenv
langchain
langchain-groq
langsmith
